Added a title for both the Days and Mood tracker pages.
