+++ lib/main.dart
@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
 import 'package:shelter_in_place/pages/home.dart';
-import 'package:shelter_in_place/pages/overview_charts.dart';
+import 'package:shelter_in_place/pages/summary/mood_tracker.dart';
 import 'package:shelter_in_place/pages/questions/note.dart';
 import 'package:shelter_in_place/pages/questions/questions_controller.dart';
 import 'package:shelter_in_place/pages/questions/social_distancing.dart';
 import 'package:shelter_in_place/pages/settings/notification_settings.dart';
 import 'package:shelter_in_place/pages/settings/user_settings.dart';
-import 'package:shelter_in_place/pages/summary/new_summary.dart';
+import 'package:shelter_in_place/pages/summary/days_tracker.dart';
 import 'package:shelter_in_place/services/backend_service.dart';
 import 'package:shelter_in_place/services/days_service.dart';
 
@@ -60,7 +60,7 @@ class MyApp extends StatelessWidget {
               future: Provider.of<AuthService>(context).getUser(),
               builder: (context, AsyncSnapshot<FirebaseUser> snapshot) {
                 if (snapshot.connectionState == ConnectionState.done) {
-                  return snapshot.hasData ? HomePage() : LoginFirstPage();
+                  return snapshot.hasData ? SocialDistancing() : LoginFirstPage();
                 } else {
                   // show loading indicator
                   return Container(color: Colors.white);
